+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+# Changed
+
+* GetGitRepoVersion now defines `GIT_DESCRIBE_HASH` and `ROQ_BUILD_TYPE`
+
 ## 0.4.2 &ndash; 2020-07-27
 
 ## 0.4.1 &ndash; 2020-07-17